+++ src/main/java/com/example/wdmsystem/category/system/ICategoryRepository.java
@@ -2,5 +2,5 @@ package com.example.wdmsystem.category.system;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICategoryRepository extends JpaRepository<Category, Long> {
+public interface ICategoryRepository extends JpaRepository<Category, Integer> {
 }
